speech/cloud-client/requirements.txt
google-cloud-speech==0.33.0
google-cloud-storage==1.7.0
